requirements.txt
pyyaml>=6.0.2,<7.0.0
feedparser>=6.0.11,<7.0.0
sentence-transformers>=3.1.0,<6.0.0
numpy>=2.0.0,<3.0.0
torch>=2.5.0,<3.0.0
ruff>=0.8.0,<1.0.0
pytest>=7.0.0,<10.0.0
